remove old main window and rename current main
